dbuslomirisessionservice: change the way we lookup our session
We previously assumed that we're under logind's session and get our
session using "GetSessionByPID" call.

First, if we actually are under a session, we'll have XDG_SESSION_ID
environment variable set and can use that. Second, we can be run in a
systemd unit, in which case we aren't under a session anymore. In this
case, ask logind for the user's "Display" session.