Wrap the text on the error screen
When Lomiri encounters an error during loading, it shows a nice error
screen to give more information without needing to console in to the
device. However, if the device's screen is too small or the text too
large, the message can overflow off the right side.

This change causes Lomiri to wrap the lines of the error message at the
screen border so it can be read in its entirety.