qml: Even faster blurred Drawer
The BackgroundBlur component used to be busy on redrawing
the blur source whenever the source rectangle changed.
This turned out to be the main reason for weaker devices
to have problems drawing with the blur shader enabled.

Instead of changing the source rect based on the Drawer's
drag distance, do the following:

- Make the BackgroundBlur stay in place attached to
  the left-hand side.
- Use a parent item with clipping enabled to cause
  the blur effect to not show up outside of the
  Drawer's boundaries.
- Change the blur rectangle to only reference the area
  of interest.

Fixes blurred Drawer performance on more devices
as tested on the JingPad, as well as the weird
jelly effect during drag which was introduced
in the last performance changeset.