qml/Stage: Fix flicker between splash & surface states
There is not much of a need to have states when there are
only two options. Just set the splashLoader's active state
to whether the respective MirSurface is ready or not.
This removes the need to be careful about it's presence.
This turns showing the app into an instant operation,
removing the need for a fade-in effect.