plugins: Avoid MirAL throwing exceptions for OOM-killed apps
When OOM-killed apps are tapped in the spread MirAL doesn't know
which app is supposed to be restarted, as the surfaces it manages
are gone by that time, and would throw an exception during access
of the map of surfaces.

In the case that the app is *really* stopped but still in the spread,
just request the app to be started during raise of the surface.
This follows the intended design of re-starting apps when they have
been killed in a mobile setting, explicitly and intentionally.