WindowManager: Fix crash during tapping of dead window surface/app
Instead of tapping on killed apps (either OOM or otherwise) causing a crash,
cause a new window to fill out the next-best window suitable as a placeholder.

QtMir in this branch changes/forces the behavior to "delete any but 1 surface",
removing them from the windowModel, but the shell needs to either remove it or
fill out the placeholder. And Lomiri per design leaves one around in case of OOM,
to tap the spread and restart it. The user needs to be able to either activate or
remove the app at any time, without the shell crashing on them.