qml: Clip SurfaceContainer contents
Since QtMir centers cropped textures now [1] but the SurfaceContainer
containing the texture doesn't crop it, entering the Spread overview
has the effect of placing the window preview outside of the boundaries
it is supposed to be.

Instead of "dancing around what the question is" clip the contents away
and show the middle of the window preview in the Spread overview.
Granted, it's supposed to be showing the header but since GTK3+4 support
*will* be a topic, work around this with clipping.

This induces a slight performance penalty on the Shell since clipping
causes a performance hit on the GPU, but this is the better solution
than anything else at this moment.

Let's treat this as a temporary workaround for now until the GTK3+4
support story really kicks into a higher gear.

[1]: https://gitlab.com/ubports/development/core/qtmir/-/merge_requests/53