systemd-user: start lal-application-end.target on stop/restart
We have a problem that SIGSTOP'ed apps don't get the memo when Lomiri
stops for any reason, causing the apps (and subsequently the LAL units)
to linger around, confusing LAL and causing those apps to not be
launch-able again.

Together with [1], make sure those apps are ended when Lomiri ends
through the use of lal-application-end.target. This restores the
functionality in the Upstart unit file.

Fixes: https://gitlab.com/ubports/development/core/lomiri-app-launch/-/issues/14

[1] https://gitlab.com/ubports/development/core/lomiri-app-launch/-/merge_requests/40