Wizard: show "Welcome to " + distro
It's silly to say "Welcome to Lomiri" since Lomiri is just 1 component
in Ubuntu Touch. Preferably this should say "Welcome to %s" to allow
translated string to move the distro around. However, I don't want to
introduce a new translatable string at this point, so reuse "Welcome
to " string from `10-welcome-update` page.

Allow the distro name to either be read from /etc/os-release, or be
configured at build time which is intended to be used for a distro
derivative which does not change /etc/os-release yet (like UBports).

Fixes: https://gitlab.com/ubports/development/core/lomiri/-/issues/57