qml: Prevent Drawer from hiding away from under the mouse pointer
When autohide is enabled the dismiss timer can snap back in the Drawer underneath
the users mouse pointer. Check whether the Drawer is pulled out or not to determine
whether to hide everything or not. Ideally, keep the Drawer dragged out until
interacted with, clicked outside of the area or hidden otherwise.