qml: Fix dragging windows in the spread getting stuck
When running in Desktop Mode on a device with a touch screen (JingPad), it turns out
to quite often happen that dragging windows in the spread overview with a finger
causes it to get stuck when switching between dragging state and hiding the app window
from the Spread's list of available surfaces.

Always cancel drags when onTouchUpdated gets called without the touch point being pressed.