data: Enhance Xwayland wrapper by setting shm mode optionally
The Pixel 3a and other Qualcomm devices with Halium 9 or lower are exposed
to driver issues in OpenGL where framebuffer objects of certain color formats
are rendered black. It's a similar symptom to [1].

To enable those devices to also gain Xwayland functionality, just allow the
maintainer of the port to specifiy whether Xwayland should run in shm mode or not.

[1]: https://gitlab.com/ubports/development/core/hybris-support/mir-android2-platform/-/merge_requests/6