cursor: Always follow cursor position from mir
It makes no sense to not, if we dont we can quickly get out of sync with
mir.

This was done for multiscreen, but this has to be fixed another way then
tracking pointer ourself. The plan is to follow mir's screen position
anyway so we don't need this. (the current multiscreen impl is horrable)