debian/session-migration: try not to overwrite settings
Due to the mistake in the mtime workaround [1], it means this script
could be run after the first boot on Focal. In that case, we must NOT
migrate the settings, as otherwise we'll overwrite existing settings.

Bug: https://gitlab.com/ubports/ubuntu-touch/-/issues/2177
(cherry picked from commit 9b9b78492dca63ca8e611646c8398b7f0337d86e)