qml/Stage: Workaround WorkspacePreview renderer crashes
Both the Stage and the WorkspacePreview present MirSurfaceItems with the
same surfaces, but having the WorkspacePreview fade in while its view is
blocked during the "Entering Spread animation" causes QSGRenderThread to
crash during QSGTexture comparisons.

This can be avoided by delaying the fade-in of the WorkspacePreview until
the toplevel MirSurfaceItems are done animating themselves.

It is not a long-term solution, rather a workaround to enable proper
functionality of workspaces without regular crashes.