qml/Shell.qml: Stop using org.gnome.desktop.background gsettings namespace for obtaining a background picture URI. Use our own namespace instead.
This allows overriding the hard-coded default background image via
system-wide and user gsettings. Note that this mostly has an effect on
initial login. Whenever the user chooses to configure a custom background
AccountsService will kick in and take over.

In previous discussions, it was argued to drop the gsettings config layer
for background images entirely and fully employ AccountsService for
configuration handling of the background image. However, gsettings is the
superior mechanism esp. for 3rd party vendors (such as distributors) to
adjust Lomiri defaults.