tst_MenuPopup: improve reliability of the test
Qt has made positioning asynchronous, which affects various tests which
rely on e.g. positioning mouse. As such, Qt has provided, in TestCase,
waitForItemPolished() which waits for positioning which happen as part
of "polishing".

However, for MenuPopup, it's important to waitForItemPolished() on the
correct object. When MenuPopup's data is updated, it's reflected in the
menuColumn which change size. However, it's contained in a Flickable
which (apparently) doesn't always change size. So we have to set an
objectName on the menuColumn and then waitForItemPolished() on it.