Revert "plugins/LightDM: fix (proxy) SessionsModel's member initialization"
As it turns out, the cause of the "wrong" devirtualization is not any
undefined behavior, but because the compilation process sees the real
liblightdm's headers which does not override roleNames().

The new code accentially loses devirtualization only because there's
more initialization before m_model->roleNames() is called, which causes
compiler to lose confidence that m_model is still the same type.

By reverting the commit, it assures that the mock behaves exactly as the
real liblightdm.

This reverts commit f4b2fa4815d77413e53e884b26f23440660ee225.