Fix compatibility with LightDM >= 1.32
liblightdm-qt >= 1.32 is no longer compatible with Qt 4 and no longer uses the
deprecated QAbstractItemModel::setRoleNames() method but rather implements
QAbstractItemModel::roleNames(). Add a compile time definition dependening on
the LightDM version and adapt the mock liblightdm-qt accordingly. This fixes
unit test failures due to the discrepancy between the original liblightdm-qt to
which the Lomiri LightDM plugin was linked with and the mocking library which
is used for the unit tests.
Also raise the required CMake version to 3.7 due to the use of VERSION_LESS.