WorkspacePreview: reduce lag when app enters/exit fullscreen
As discovered in #122, after commit c4f50475 ("Fixes and adjustments in
the workspace previews") WorkspacePreview is being resized when an app
enter or leave fullscreen mode. This causes preview's background image
to be reloaded as well.

Commit 3f991507 ("Workspaces: use AvailableDesktopArea for screen space
instead") allows image reload to not be tied to animation. This commit
additionally set image to load asynchronously, which should reduce lag
further.

Bug: https://gitlab.com/ubports/development/core/lomiri/-/issues/122