plugins & qml: Fix moving mouse cursor to second screen
The Cursor plugin implements Lomiri-specific logic for detecting edge gestures.
Due to the way it handled active and current screens with pointers it was having
a discrepancy between what QML believed was active and C++ believed was the current
pointer screen.

Simplify this by providing the "active"-ness of MousePointers from QML with a
property, and have the screen detection logic take over in C++.
Set the "enabled" property of the MousePointer from the InputDispatcherFilter
similarly to how previously it was handled in QML, to only show a single
mouse pointer on the currently used screen.

The mouse pointer can now flow between screens as expected.