GestureAreaSizeHint: don't display on height change, including on rotate
Despite the description of MR 201 [1] says, the visual hint will also be
displayed when height is changed, not only width. This is problematic;
the height of the area will be changed when the shell is rotated,
resulting in the hint being displayed in a seemingly unrelated
situation.

So just remove the handler for height change. The usecase of this seems
to be to display visually when the width is adjusted in the settings;
this commit should not affect that.

Fixes: 08d48e53 ("Add visual hint when left and right edge gesture areas are resized")

[1]: https://gitlab.com/ubports/development/core/lomiri/-/merge_requests/201