plugins/Lomiri/Launcher: remove "leaked" item from list before deleting
Since the "leaked" item is already added to the list, one can't simply
delete it. It has to be removed from the list first, or it will be a
use-after-free.

Actually, it's not clear under what situation will the item be added but
not returned from `findApplication()` (as ASAdapter doesn't seem to
modify the list). But I'll let that code stay...

Fixes: 1430cd6d ("plugins/Lomiri/Launcher: Plug a memory leak")