Workspaces: use AvailableDesktopArea for screen space instead
We already have AvailableDesktopArea which calculate this, and it's
already plumbed into Stage for another use. So also plumb it to
Workspaces for, well, available desktop area instead of calculating
its own.

But the visible benefit of this change is that, because the values used
by AvailableDesktopArea is not subject to animation, we don't have to
animate the resize of WorkspacePreview needlessly. Because it's
expensive to resize WorkspacePreview (because it also re-load the
background image), this in turn reduces lag when entering and leaving
fullscreen mode [1]. There's still a pause at the beginning of
transition, but this is already a lot better than before.

[1]: https://gitlab.com/ubports/development/core/lomiri/-/issues/122

Bug: https://gitlab.com/ubports/development/core/lomiri/-/issues/122