qml/, tests/: follow rename { => lomiri-}dialer-app
Lomiri hardcodes app id of Dialer in various places, for handling of
calls in various states (e.g. when incoming call arrives when locked, or
when requesting emergency call on the lock screen). Update them in all
relevant places.

Hardcoded app ids in tests are also updated. Now, it so happen that
Dialer app is also used generically in the test (as an ordinary mocked
app), which makes this commit larger than otherwise.