InputDispatchFilter: take Shell's rotation into account
It's incorrect to simply add the screen's top left into MousePointer's
position to turn it into a global position; one must use QQuickItem::
mapToGlobal() to let QtQuick take rotation (among others) into account.
Since the position is relative to the item's parent, call to its parent
for this.

On a related note, also map relative movement into global cordinate
using the same method, as the relative movement is relative to post-
rotate cursor position, not to global position.

Now, from testing it seems like this codepath causes an issue on some
mainline devices, so gate this behind a feature flag.

Bug: https://gitlab.com/ubports/development/core/lomiri/-/work_items/79