plugins/WindowManager: fix multi-process app having phantom window
If an application has 2 sessions (i.e. has 2 processes), and one of them
exits, then by the time TLWM receives `onSurfaceDied()`, the surface
would already be removed from application's surface list. Thus, for the
purpose of determining whether we should retain a placeholder window, we
cannot rely on that.

Instead, just check against our own list of windows to count windows for
a particular application, and use that instead of application's surface
list.