Wizard/10-welcome-update: show version from system-image tag
We used to show version from current framework, but now framework will
be fixed per major release, resulting in string like "Welcome to Ubuntu
Touch 24.04-1.x".

So, instead, pull version from SystemImage version tag, which will
contain the actual version of the current image. If the image doesn't
contain a tag, then we pull it from current framework as a failback.

Note: I know that this is completely UT-specific. However, for now the
wizard itself is (AFAICT) completely broken outside of UT anyway.
Changes to make Lomiri wizard more distro-agnostic in the future are
welcomed...

Fixes: https://gitlab.com/ubports/development/core/lomiri/-/issues/135