qml: Avoid breaking bindings of the LauncherPanel's "visible" property
The LauncherPanel's peekingAnimation used to set it's root "visible" property
during animations using PropertyActions, breaking bindings in the Launcher's
use of the "visible" property. This was especially apparent in full-shell
mode where the LauncherPanel would disappear when a launcher icon has
an alert animation happening.

Instead, create a readonly property "hasPeekingIcon" which tells the Launcher
to set the panel back to visible again as part of the binding.